Package common functionality into beam-backend-bench
