+++ .gitignore
@@ -25,8 +25,3 @@ docs/ChinookData
 result
 result-*
 .direnv
-# Profiling output
-*.prof
-*.eventlog
-*.eventlog.html
-*.hp
