+++ beam-postgres/beam-postgres.cabal
@@ -102,26 +102,6 @@ test-suite beam-postgres-tests
                       ScopedTypeVariables, MultiParamTypeClasses, TypeApplications, DeriveGeneric,
                       DeriveAnyClass, RankNTypes
 
-benchmark beam-postgres-bench
-  type:           exitcode-stdio-1.0
-  hs-source-dirs: bench
-  main-is:        Main.hs
-  other-modules:  BenchSchema
-  build-depends:
-    base,
-    beam-core,
-    beam-postgres,
-    bytestring,
-    deepseq,
-    postgresql-simple,
-    split,
-    tasty,
-    tasty-bench,
-    testcontainers,
-    text
-  default-language:   Haskell2010
-  ghc-options:        -Wall -O2 -rtsopts
-
 flag werror
   description: Enable -Werror during development
   default:     False
