+++ beam-sqlite/beam-sqlite.cabal
@@ -94,24 +94,6 @@ test-suite beam-sqlite-tests
     TypeApplications,
     TypeFamilies
 
-benchmark beam-sqlite-bench
-  type:           exitcode-stdio-1.0
-  hs-source-dirs: bench
-  main-is:        Main.hs
-  other-modules:  BenchSchema
-  build-depends:
-    base,
-    beam-core,
-    beam-sqlite,
-    deepseq,
-    split,
-    sqlite-simple,
-    tasty,
-    tasty-bench,
-    text
-  default-language:   Haskell2010
-  ghc-options:        -Wall -O2 -rtsopts
-
 flag werror
   description: Enable -Werror during development
   default:     False
