--- cabal.project.profiling
@@ -1,4 +1,0 @@
-import: cabal.project
-profiling: True
-executable-profiling: True
-profiling-detail: all-functions
